fix: app image, 스켈레톤 위젯 분리 리팩토링
