+++ lib/presentation/component/app_image.dart
@@ -9,6 +9,8 @@ class AppImage extends StatefulWidget {
   final BorderRadius borderRadius;
   final VoidCallback? onImageLoaded;
   final Widget Function(BuildContext context)? overlayBuilder; //오버레이 넣는 위젯 부분
+  final Widget Function(BuildContext context, Size size)?
+  skeletonBuilder; //스켈레톤 넣는 위젯 부분
 
   const AppImage({
     super.key,
@@ -19,6 +21,7 @@ class AppImage extends StatefulWidget {
     this.borderRadius = BorderRadius.zero,
     this.onImageLoaded,
     this.overlayBuilder,
+    this.skeletonBuilder,
   });
 
   @override
@@ -31,6 +34,10 @@ class _AppImageState extends State<AppImage> {
   @override
   Widget build(BuildContext context) {
     final isNetwork = widget.path.startsWith('http');
+    final imageSize = Size(
+      widget.width ?? double.infinity,
+      widget.height ?? double.infinity,
+    );
 
     if (isNetwork) {
       return ClipRRect(
@@ -57,7 +64,9 @@ class _AppImageState extends State<AppImage> {
               },
               errorBuilder: (_, __, ___) => _errorPlaceholder(),
             ),
-            if (!_loaded) _skeletonAnimation(),
+            if (!_loaded)
+              widget.skeletonBuilder?.call(context, imageSize) ??
+                  const SizedBox(),
             if (_loaded && widget.overlayBuilder != null)
               widget.overlayBuilder!(context),
           ],
@@ -88,83 +97,4 @@ class _AppImageState extends State<AppImage> {
     height: widget.height,
     child: const Icon(Icons.photo, color: ColorStyle.black),
   );
-
-  Widget _skeletonAnimation() {
-    return SkeletonAnimationWidget(width: widget.width, height: widget.height);
-  }
-}
-
-class SkeletonAnimationWidget extends StatefulWidget {
-  final double? width;
-  final double? height;
-
-  const SkeletonAnimationWidget({super.key, this.width, this.height});
-
-  @override
-  State<SkeletonAnimationWidget> createState() =>
-      _SkeletonAnimationWidgetState();
-}
-
-class _SkeletonAnimationWidgetState extends State<SkeletonAnimationWidget>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<double> _animation;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 1500),
-    );
-
-    _animation = Tween<double>(begin: -2, end: 2).animate(
-      CurvedAnimation(parent: _controller, curve: Curves.easeInOutSine),
-    );
-
-    _controller.repeat(reverse: true);
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: widget.width,
-      height: widget.height,
-      decoration: const BoxDecoration(color: ColorStyle.primary60),
-      child: AnimatedBuilder(
-        animation: _animation,
-        builder: (_, __) {
-          return ShaderMask(
-            shaderCallback: (bounds) {
-              return LinearGradient(
-                begin: Alignment.topLeft,
-                end: Alignment.bottomRight,
-                colors: const [
-                  ColorStyle.primary100,
-                  ColorStyle.primary60,
-                  ColorStyle.primary80,
-                ],
-                stops: [
-                  0.0,
-                  (_animation.value + 2) / 4, // 0.0 ~ 1.0 범위로 정규화
-                  1.0,
-                ],
-              ).createShader(bounds);
-            },
-            child: Container(
-              width: widget.width,
-              height: widget.height,
-              color: ColorStyle.primary60,
-            ),
-          );
-        },
-      ),
-    );
-  }
 }
